moved to services; renamed class; return a Set to ensure no duplicates
